requirements.txt
Cython==0.29.33
fvcore==0.1.5.post20221221
imageio==2.25.0
ipdb==0.13.11
matplotlib==3.6.3
natsort==8.2.0
numpy==1.23.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
opencv-python==4.7.0.68
Pillow==9.4.0
prettytable==3.6.0
pydicom==2.3.0
pyrender
pytorch-lightning==1.9.0
pyvista==0.36.1
PyYAML==6.0
scikit-image==0.19.3
scipy==1.10.0
smplpytorch==0.0.8
smplx==0.1.28
tensorboard==2.11.2
torch==1.13.1
torchaudio==0.13.1
torchmetrics==0.11.0
torchvision==0.14.1
tqdm==4.64.1
trimesh==3.18.1
wandb==0.13.9
numba